Integrated Pippin's menu template and got a good start on game one.
Had to modify a bunch of code to get the menu template to work, but now it does and it's pretty clean. First game almost done. In this one, you have to catch all the flies to win. I also added an a screen which pops up when you win. I'll probably reuse it in the other games. Also added a cool font!