Variation 2 is complete!
This one is very similar to the first variation but instead of having as much time as you want to eat all 200 flies, you have 5 seconds. Other cool things i did in this variation: - create a seconds variable to countdown from 5 seconds, created a variable called gameIsGaming (when false, there is no more tongue and tongue overlap), created a function for countdown, created a second ending (if only caught 1 fly).